Moved most utils from framework/utils to lib/utils
