+++ frontend/src/framework/internal/components/Content/private-components/ViewWrapper/viewWrapper.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import { ModuleInstance } from "@framework/ModuleInstance";
 import { Workbench } from "@framework/Workbench";
-import { Point, pointRelativeToDomRect, pointerEventToPoint, pointDifference } from "@framework/utils/geometry";
+import { Point, pointDifference, pointRelativeToDomRect, pointerEventToPoint } from "@lib/utils/geometry";
 
 import { Header } from "./private-components/header";
 import { ViewContent } from "./private-components/viewContent";
