+++ frontend/src/lib/utils/geometry.ts
